Moved code from file_common
